Quick fix to the delete users when they have open applications

+++ src/main/java/com/park/easyrecruit/ejb/UserBean.java
@@ -136,13 +136,20 @@ public class UserBean {
         user.setPosition(position);
     }
 
-    public void deleteUsers(List<Integer> ids) {
+    public boolean deleteUsers(List<Integer> ids) {
         LOG.info("UserBean:deleteUserByIds");
 
         for (Integer id : ids) {
             User user = em.find(User.class, id);
+            
+            if (user.getApplications().size() > 0) {
+                return false;
+            }
+            
             em.remove(user);
         }
+        
+        return true;
     }
 
     public void deleteUser(Integer userId) {
